Move deployment info struct.
